+++ model_utils/fields.py
@@ -1,8 +1,10 @@
 from __future__ import unicode_literals
 
 import django
+import uuid
 from django.db import models
 from django.conf import settings
+from django.core.exceptions import ValidationError
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.timezone import now
 
@@ -17,6 +19,7 @@ class AutoCreatedField(models.DateTimeField):
     By default, sets editable=False, default=datetime.now.
 
     """
+
     def __init__(self, *args, **kwargs):
         kwargs.setdefault('editable', False)
         kwargs.setdefault('default', now)
@@ -30,6 +33,7 @@ class AutoLastModifiedField(AutoCreatedField):
     By default, sets editable=False and default=datetime.now.
 
     """
+
     def pre_save(self, model_instance, add):
         value = now()
         if not model_instance.pk:
@@ -53,6 +57,7 @@ class StatusField(models.CharField):
     Also features a ``no_check_for_status`` argument to make sure
     South can handle this field when it freezes a model.
     """
+
     def __init__(self, *args, **kwargs):
         kwargs.setdefault('max_length', 100)
         self.check_for_status = not kwargs.pop('no_check_for_status', False)
@@ -93,6 +98,7 @@ class MonitorField(models.DateTimeField):
     changes.
 
     """
+
     def __init__(self, *args, **kwargs):
         kwargs.setdefault('default', now)
         monitor = kwargs.pop('monitor', None)
@@ -144,7 +150,9 @@ SPLIT_MARKER = getattr(settings, 'SPLIT_MARKER', '<!-- split -->')
 # the number of paragraphs after which to split if no marker
 SPLIT_DEFAULT_PARAGRAPHS = getattr(settings, 'SPLIT_DEFAULT_PARAGRAPHS', 2)
 
-_excerpt_field_name = lambda name: '_%s_excerpt' % name
+
+def _excerpt_field_name(name):
+    return '_%s_excerpt' % name
 
 
 def get_excerpt(content):
@@ -252,3 +260,49 @@ class SplitField(models.TextField):
         name, path, args, kwargs = super(SplitField, self).deconstruct()
         kwargs['no_excerpt_field'] = True
         return name, path, args, kwargs
+
+
+class UUIDField(models.UUIDField):
+    """
+    A field for storing universally unique identifiers. Use Python UUID class.
+    """
+
+    def __init__(self, primary_key=True, version=4, editable=False, *args, **kwargs):
+        """
+        Parameters
+        ----------
+        primary_key : bool
+            If True, this field is the primary key for the model.
+        version : int
+            An integer that set default UUID version.
+        editable : bool
+            If False, the field will not be displayed in the admin or any other ModelForm,
+            default is false.
+
+        Raises
+        ------
+        ValidationError
+            UUID version 2 is not supported.
+        """
+        
+        if version == 2:
+            raise ValidationError(
+                'UUID version 2 is not supported.')
+            
+        if version < 1 or version > 5:
+            raise ValidationError(
+                'UUID version is not valid.')
+            
+        if version == 1:
+            default = uuid.uuid1
+        elif version == 3:
+            default = uuid.uuid3
+        elif version == 4:
+            default = uuid.uuid4
+        elif version == 5:
+            default = uuid.uuid5  
+           
+        kwargs.setdefault('primary_key', primary_key)
+        kwargs.setdefault('editable', editable)
+        kwargs.setdefault('default', default)
+        super(UUIDField, self).__init__(*args, **kwargs)