+++ CHANGES.rst
@@ -17,6 +17,7 @@ master (unreleased)
 - Fix patched `save` in FieldTracker
 - Upgrades test requirements (pytest, pytest-django, pytest-cov) and
   skips tox test with Python 3.5 and Django (trunk)
+- Add UUIDModel and UUIDField support.
 
 3.1.2 (2018.05.09)
 ------------------